requirements.txt
Brotli==1.0.9
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.4
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.3
cycler==0.12.1
filelock==3.13.1
fonttools==4.45.1
idna==3.4
importlib-resources==6.1.1
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.1
matplotlib==3.7.4
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==6.2.2
pip==23.3.1
pycparser==2.21
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
setuptools==68.0.0
six==1.16.0
sympy==1.11.1
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==1.26.18
wheel==0.41.2
win-inet-pton==1.1.0
zipp==3.17.0
